requirements.txt
torch~=1.9.0+cu111
torchvision~=0.10.0+cu111
opencv-python~=4.4.0.46
Pillow~=9.0.1
elasticsearch~=8.1.2
mxnet~=1.9.0
PyYAML~=6.0
tqdm~=4.64.0
matplotlib~=3.5.1
numpy~=1.22.3
pandas~=1.4.2
seaborn~=0.11.2
requests~=2.27.1
Flask~=2.1.1
thop~=0.0.31.post2005241907
easydict~=1.9
imageio~=2.16.1
sklearn~=0.0
scikit-learn~=0.23.2
scipy~=1.8.0
six~=1.16.0
onnx~=1.9.0
onnxruntime~=1.11.0
scikit-image~=0.19.2
setuptools~=58.0.4
prettytable~=3.2.0
visualdl~=2.2.3
trimesh~=3.10.7